# openapi/registry.openapi.json
{
  "components": {
    "schemas": {
      "ActorReference": {
        "additionalProperties": false,
        "properties": {
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "actor_type": {
            "$ref": "#/components/schemas/ActorType"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "required": [
          "actor_id",
          "actor_type"
        ],
        "title": "ActorReference",
        "type": "object"
      },
      "ActorType": {
        "enum": [
          "human",
          "system",
          "tool",
          "organization"
        ],
        "title": "ActorType",
        "type": "string"
      },
      "Adjudication": {
        "additionalProperties": false,
        "properties": {
          "adjudication_id": {
            "pattern": "^oc[a-z]*_[A-Za-z0-9._-]+$",
            "title": "Adjudication Id",
            "type": "string"
          },
          "adjudication_round": {
            "minimum": 1.0,
            "title": "Adjudication Round",
            "type": "integer"
          },
          "adjudicator_id": {
            "title": "Adjudicator Id",
            "type": "string"
          },
          "anchors_reviewed": {
            "title": "Anchors Reviewed",
            "type": "boolean"
          },
          "blinded_fields": {
            "items": {
              "type": "string"
            },
            "title": "Blinded Fields",
            "type": "array"
          },
          "concern_id": {
            "pattern": "^oc[a-z]*_[A-Za-z0-9._-]+$",
            "title": "Concern Id",
            "type": "string"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "conflict_declaration": {
            "$ref": "#/components/schemas/ConflictDeclaration"
          },
          "content_hash": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Content Hash",
            "type": "string"
          },
          "counterposition_assessment": {
            "title": "Counterposition Assessment",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "$ref": "#/components/schemas/ActorReference"
          },
          "evidence_ids": {
            "items": {
              "pattern": "^oc[a-z]*_[A-Za-z0-9._-]+$",
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "id": {
            "pattern": "^oc[a-z]*_[A-Za-z0-9._-]+$",
            "title": "Id",
            "type": "string"
          },
          "reasoning": {
            "minLength": 1,
            "title": "Reasoning",
            "type": "string"
          },
          "requested_followup": {
            "items": {
              "$ref": "#/components/schemas/FollowupRequest"
            },
            "title": "Requested Followup",
            "type": "array"
          },
          "schema_version": {
            "default": "0.1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "validity": {
            "$ref": "#/components/schemas/ValidityDecision"
          }
        },
        "required": [
          "id",
          "created_by",
          "content_hash",
          "adjudication_id",
          "concern_id",
          "adjudicator_id",
          "adjudication_round",
          "blinded_fields",
          "conflict_declaration",
          "validity",
          "severity",
          "confidence",
          "reasoning",
          "evidence_ids",
          "counterposition_assessment",
          "anchors_reviewed"
        ],
        "title": "Adjudication",
        "type": "object"
      },
      "AdjudicationSubmissionInput": {
        "additionalProperties": false,
        "properties": {
          "anchors_reviewed": {
            "title": "Anchors Reviewed",
            "type": "boolean"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "conflict_declaration": {
            "$ref": "#/components/schemas/ConflictDeclaration"
          },
          "counterposition_assessment": {
            "minLength": 1,
            "title": "Counterposition Assessment",
            "type": "string"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "reasoning": {
            "minLength": 20,
            "title": "Reasoning",
            "type": "string"
          },
          "requested_followup": {
            "items": {
              "$ref": "#/components/schemas/FollowupRequest"
            },
            "title": "Requested Followup",
            "type": "array"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "validity": {
            "$ref": "#/components/schemas/ValidityDecision"
          }
        },
        "required": [
          "validity",
          "severity",
          "confidence",
          "reasoning",
          "counterposition_assessment",
          "anchors_reviewed",
          "conflict_declaration"
        ],
        "title": "AdjudicationSubmissionInput",
        "type": "object"
      },
...
      "AnchorType": {
        "enum": [
          "text_span",
          "equation",
          "theorem",
          "figure",
          "panel",
          "table",
          "cell",
          "citation",
          "bibliography_entry",
          "code_reference",
          "section"
        ],
        "title": "AnchorType",
        "type": "string"
      },
...
      "BlindedAnchor": {
        "additionalProperties": false,
        "properties": {
          "anchor_id": {
            "title": "Anchor Id",
            "type": "string"
          },
          "anchor_type": {
            "$ref": "#/components/schemas/AnchorType"
          },
          "bounding_boxes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Bounding Boxes",
            "type": "array"
          },
          "extraction_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Confidence"
          },
          "normalized_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Text"
          },
          "object_coordinates": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                }
              ]
            },
            "title": "Object Coordinates",
            "type": "object"
          },
          "object_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Label"
          },
          "page_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page End"
          },
          "page_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Start"
          },
          "rendered_artifact_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rendered Artifact Sha256"
          },
          "resolution_status": {
            "title": "Resolution Status",
            "type": "string"
          },
          "section_path": {
            "items": {
              "type": "string"
            },
            "title": "Section Path",
            "type": "array"
          },
          "source_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Text"
          }
        },
        "required": [
          "anchor_id",
          "anchor_type",
          "page_start",
          "page_end",
          "bounding_boxes",
          "section_path",
          "source_text",
          "normalized_text",
          "object_label",
          "object_coordinates",
          "extraction_confidence",
          "rendered_artifact_sha256",
          "resolution_status"
        ],
        "title": "BlindedAnchor",
        "type": "object"
      },
      "BlindedClaim": {
        "additionalProperties": false,
        "properties": {
          "anchor_ids": {
            "items": {
              "type": "string"
            },
            "title": "Anchor Ids",
            "type": "array"
          },
          "approval_status": {
            "title": "Approval Status",
            "type": "string"
          },
          "claim_id": {
            "title": "Claim Id",
            "type": "string"
          },
          "claim_type": {
            "title": "Claim Type",
            "type": "string"
          },
          "explicitness": {
            "title": "Explicitness",
            "type": "string"
          },
          "reconstruction_notes": {
            "title": "Reconstruction Notes",
            "type": "string"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "statement": {
            "title": "Statement",
            "type": "string"
          }
        },
        "required": [
          "claim_id",
          "statement",
          "claim_type",
          "explicitness",
          "scope",
          "anchor_ids",
          "reconstruction_notes",
          "approval_status"
        ],
        "title": "BlindedClaim",
        "type": "object"
      },
      "BlindedCounterposition": {
        "additionalProperties": false,
        "properties": {
          "adequacy_status": {
            "title": "Adequacy Status",
            "type": "string"
          },
          "counterposition_id": {
            "title": "Counterposition Id",
            "type": "string"
          },
          "residual_disagreement": {
            "title": "Residual Disagreement",
            "type": "string"
          },
          "source_label": {
            "const": "blinded_counterposition",
            "default": "blinded_counterposition",
            "title": "Source Label",
            "type": "string"
          },
          "statement": {
            "title": "Statement",
            "type": "string"
          },
          "supporting_anchor_ids": {
            "items": {
              "type": "string"
            },
            "title": "Supporting Anchor Ids",
            "type": "array"
          },
          "supporting_evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Supporting Evidence Ids",
            "type": "array"
          }
        },
        "required": [
          "counterposition_id",
          "statement",
          "supporting_anchor_ids",
          "supporting_evidence_ids",
          "residual_disagreement",
          "adequacy_status"
        ],
        "title": "BlindedCounterposition",
        "type": "object"
      },
      "BlindedEvidence": {
        "additionalProperties": false,
        "properties": {
          "anchor_ids": {
            "items": {
              "type": "string"
            },
            "title": "Anchor Ids",
            "type": "array"
          },
          "artifact_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Sha256"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "evidence_type": {
            "title": "Evidence Type",
            "type": "string"
          },
          "independence_group": {
            "title": "Independence Group",
            "type": "string"
          },
          "limitations": {
            "title": "Limitations",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "reproducibility_status": {
            "title": "Reproducibility Status",
            "type": "string"
          },
          "source_label": {
            "const": "blinded_source",
            "default": "blinded_source",
            "title": "Source Label",
            "type": "string"
          },
          "supports": {
            "title": "Supports",
            "type": "string"
          }
        },
        "required": [
          "evidence_id",
          "evidence_type",
          "supports",
          "description",
          "anchor_ids",
          "method",
          "reproducibility_status",
          "limitations",
          "independence_group",
          "artifact_sha256"
        ],
        "title": "BlindedEvidence",
        "type": "object"
      },
      "BlindedTaskPayload": {
        "additionalProperties": false,
        "properties": {
          "anchors": {
            "items": {
              "$ref": "#/components/schemas/BlindedAnchor"
            },
            "title": "Anchors",
            "type": "array"
          },
          "blinded_fields": {
            "items": {
              "type": "string"
            },
            "title": "Blinded Fields",
            "type": "array"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "case_type": {
            "title": "Case Type",
            "type": "string"
          },
          "case_version": {
            "title": "Case Version",
            "type": "string"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/BlindedClaim"
            },
            "title": "Claims",
            "type": "array"
          },
          "concern_id": {
            "title": "Concern Id",
            "type": "string"
          },
          "concern_summary": {
            "title": "Concern Summary",
            "type": "string"
          },
          "concern_title": {
            "title": "Concern Title",
            "type": "string"
          },
          "concern_type": {
            "title": "Concern Type",
            "type": "string"
          },
          "counterpositions": {
            "items": {
              "$ref": "#/components/schemas/BlindedCounterposition"
            },
            "title": "Counterpositions",
            "type": "array"
          },
          "domain_profile": {
            "title": "Domain Profile",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/BlindedEvidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "manuscript_artifact_sha256": {
            "title": "Manuscript Artifact Sha256",
            "type": "string"
          },
          "manuscript_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manuscript Title"
          },
          "prior_adjudications": {
            "items": {
              "$ref": "#/components/schemas/PriorAdjudicationView"
            },
            "title": "Prior Adjudications",
            "type": "array"
          },
          "proposed_consequence": {
            "title": "Proposed Consequence",
            "type": "string"
          },
          "proposed_resolution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Resolution"
          },
          "rendered_artifact_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rendered Artifact Sha256"
          },
          "slot": {
            "$ref": "#/components/schemas/TaskSlot"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "uncertainty_sources": {
            "items": {
              "$ref": "#/components/schemas/UncertaintySource"
            },
            "title": "Uncertainty Sources",
            "type": "array"
          }
        },
        "required": [
          "task_id",
          "slot",
          "case_id",
          "case_version",
          "case_type",
          "manuscript_title",
          "domain_profile",
          "language",
          "manuscript_artifact_sha256",
          "rendered_artifact_sha256",
          "concern_id",
          "concern_title",
          "concern_summary",
          "concern_type",
          "proposed_consequence",
          "proposed_resolution",
          "uncertainty_sources",
          "claims",
          "anchors",
          "evidence",
          "counterpositions",
          "blinded_fields"
        ],
        "title": "BlindedTaskPayload",
        "type": "object"
      },
...
      "ConflictDeclaration": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "status": {
            "enum": [
              "none",
              "disclosed",
              "disqualifying"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "ConflictDeclaration",
        "type": "object"
      },
...
      "FollowupRequest": {
        "additionalProperties": false,
        "properties": {
          "blocking": {
            "default": false,
            "title": "Blocking",
            "type": "boolean"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "request_type": {
            "title": "Request Type",
            "type": "string"
          }
        },
        "required": [
          "request_type",
          "description"
        ],
        "title": "FollowupRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PriorAdjudicationView": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "counterposition_assessment": {
            "title": "Counterposition Assessment",
            "type": "string"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "reasoning": {
            "title": "Reasoning",
            "type": "string"
          },
          "requested_followup": {
            "items": {
              "$ref": "#/components/schemas/FollowupRequest"
            },
            "title": "Requested Followup",
            "type": "array"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "slot": {
            "$ref": "#/components/schemas/TaskSlot"
          },
          "validity": {
            "$ref": "#/components/schemas/ValidityDecision"
          }
        },
        "required": [
          "slot",
          "validity",
          "severity",
          "confidence",
          "reasoning",
          "evidence_ids",
          "counterposition_assessment",
          "requested_followup"
        ],
        "title": "PriorAdjudicationView",
        "type": "object"
      },
...
      "Severity": {
        "enum": [
          "critical",
          "major",
          "moderate",
          "minor",
          "informational"
        ],
        "title": "Severity",
        "type": "string"
      },
...
      "SubmissionView": {
        "additionalProperties": false,
        "properties": {
          "adjudication": {
            "$ref": "#/components/schemas/Adjudication"
          },
          "task": {
            "$ref": "#/components/schemas/TaskView"
          }
        },
        "required": [
          "adjudication",
          "task"
        ],
        "title": "SubmissionView",
        "type": "object"
      },
...
      "TaskSlot": {
        "enum": [
          "primary",
          "secondary",
          "tie_break"
        ],
        "title": "TaskSlot",
        "type": "string"
      },
      "TaskStatus": {
        "enum": [
          "pending",
          "claimed",
          "completed",
          "cancelled"
        ],
        "title": "TaskStatus",
        "type": "string"
      },
      "TaskView": {
        "additionalProperties": false,
        "properties": {
          "assigned_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "case_version": {
            "title": "Case Version",
            "type": "string"
          },
          "claimed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimed At"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "concern_id": {
            "title": "Concern Id",
            "type": "string"
          },
          "slot": {
            "$ref": "#/components/schemas/TaskSlot"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "case_id",
          "case_version",
          "concern_id",
          "slot",
          "status",
          "assigned_to",
          "claimed_at",
          "completed_at"
        ],
        "title": "TaskView",
        "type": "object"
      },
...
      "UncertaintySource": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "description"
        ],
        "title": "UncertaintySource",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidityDecision": {
        "enum": [
          "confirmed",
          "qualified",
          "rejected",
          "unresolved"
        ],
        "title": "ValidityDecision",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Immutable case registry, granular data-use authorization, and blinded scientific concern adjudication, expert calibration, rights-cleared intake, and claim reconstruction.",
    "title": "OpenCritique Registry and Adjudication API",
    "version": "0.5.0-alpha"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/tasks/{task_id}": {
      "get": {
        "operationId": "task_payload_v1_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlindedTaskPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Task Payload"
      }
    },
    "/v1/tasks/{task_id}/submit": {
      "post": {
        "operationId": "submit_task_v1_tasks__task_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjudicationSubmissionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Task"
      }
    },
...
  }
}